Refactoring specs + adding one for failures
