+++ Specifications/Kernel/Grains/Observation/for_CatchUp/given/a_catch_up_worker.cs
@@ -4,6 +4,7 @@
 using System.Text.Json;
 using Aksio.Cratis.EventSequences;
 using Aksio.Cratis.Execution;
+using Aksio.Cratis.Specifications;
 using Microsoft.Extensions.Logging;
 using Orleans;
 using Orleans.Runtime;
@@ -27,6 +28,8 @@ public class a_catch_up_worker : GrainSpecification
         Guid.NewGuid(),
         Guid.NewGuid(),
         Guid.NewGuid());
+
+    protected virtual IEnumerable<AppendedEvent> events => Enumerable.Empty<AppendedEvent>();
 
     protected override Grain GetGrainInstance()
     {
@@ -62,5 +65,17 @@ public class a_catch_up_worker : GrainSpecification
     {
         grain_factory.Setup(_ => _.GetGrain(typeof(ObserverSubscriber), state.ObserverId, IsAny<string>())).Returns(subscriber.Object);
         grain_factory.Setup(_ => _.GetGrain<IObserverSupervisor>(IsAny<Guid>(), IsAny<string>(), IsAny<string>())).Returns(supervisor.Object);
+
+        event_sequence_storage_provider
+            .Setup(_ => _.GetFromSequenceNumber(IsAny<EventSequenceId>(), IsAny<EventSequenceNumber>(), IsAny<EventSourceId>(), IsAny<IEnumerable<EventType>>()))
+            .Returns(() => Task.FromResult<IEventCursor>(new EventCursorForSpecifications(events)));
+
+        timer_registry
+            .Setup(_ => _.RegisterTimer(grain, IsAny<Func<object, Task>>(), IsAny<object>(), IsAny<TimeSpan>(), IsAny<TimeSpan>()))
+            .Returns((Grain __, Func<object, Task> callback, object state, TimeSpan ___, TimeSpan ____) =>
+            {
+                callback(state);
+                return Task.CompletedTask;
+            });
     }
 }
